pyproject.toml
[tool.poetry]
name = "vignette"
version = "0.1.0"
description = "Dr. Meehl's Website for Pilots and Aircraft"
authors = ["Ian Kollipara <ian.kollipara@cune.org>", "Neal Patron <patron.neal@gmail.com>"]
license = "BSD-2 Clause"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.85.0"
tortoise-orm = {extras = ["aiosqlite"], version = "^0.19.2"}

[tool.poetry.dev-dependencies]
black = "^22.8.0"
pytest = "^7.1.3"
uvicorn = "^0.18.3"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
